commandes-service/requirements.txt
SQLAlchemy==1.4.47
anyio==4.9.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
colorama==0.4.6
coverage==7.8.0
cryptography==45.0.4
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.2.0
uvicorn==0.18.3
fastapi==0.110.3
greenlet==3.1.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.1
idna==3.10
iniconfig==2.1.0
packaging==24.2
pika==1.3.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==1.10.14
pytest==8.3.5
pytest-cov==6.1.1
python-dotenv==1.0.1
python-jose[cryptography]==3.5.0
rfc3986==1.5.0
rsa==4.9.1
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.37.2
typing-extensions==4.13.2
Faker
prometheus-client
starlette-exporter
